Split multiplayer code into its own class
