+++ src/citra_qt/multiplayer/direct_connect.cpp
@@ -12,6 +12,7 @@
 #include "citra_qt/multiplayer/client_room.h"
 #include "citra_qt/multiplayer/direct_connect.h"
 #include "citra_qt/multiplayer/message.h"
+#include "citra_qt/multiplayer/state.h"
 #include "citra_qt/multiplayer/validation.h"
 #include "citra_qt/ui_settings.h"
 #include "core/settings.h"
@@ -124,7 +125,7 @@ void DirectConnectWindow::OnConnection() {
             ShowError(NetworkMessage::USERNAME_IN_USE);
             break;
         case Network::RoomMember::State::Joining:
-            auto parent = static_cast<GMainWindow*>(parentWidget());
+            auto parent = static_cast<MultiplayerState*>(parentWidget());
             parent->OnOpenNetworkRoom();
             close();
         }
